Extract commom methods in an abstraction
