+++ app/services/movie/genres.rb
@@ -1,37 +1,17 @@
-require 'faraday'
+require_relative './base.rb'
 
 module Services
   module Movie
-    class Genres
+    class Genres < Base
 
-      def initialize base_url, api_key
-        @base_url = base_url
-        @api_key = api_key
-      end
-
-      def get
-        check_arguments
-        endpoint = build_endpoint
-        result = Faraday.get endpoint
-        parsed = parse_results result
-      end
-
-      private
+      protected
 
       def path
         '/genre/movie/list'
       end
 
-      def build_endpoint
-        "#{@base_url}#{path}?api_key=#{@api_key}&language=en-US"
-      end
-
-      def check_arguments
-        raise ArgumentError, 'Param missing! Checkout BASE_URL and API_KEY env variables' unless @base_url or @api_key
-      end
-
-      def parse_results result
-        JSON.parse(result.body)['genres']
+      def handle_results
+        @parsed_results['genres']
       end
 
     end
